+++ lua/autorun/server/ttt2_force_download.lua
@@ -113,7 +113,12 @@ resource.AddFile("materials/vgui/ttt/perks/hud_noenergydmg.png")
 resource.AddFile("materials/vgui/ttt/icon_nohazarddmg.vmt")
 resource.AddFile("materials/vgui/ttt/perks/hud_nohazarddmg.png")
 
+resource.AddFile("materials/vgui/ttt/icon_beacon.vmt")
+resource.AddFile("materials/vgui/ttt/icon_confgrenade.vmt")
+resource.AddFile("materials/vgui/ttt/icon_decoy.vmt")
+resource.AddFile("materials/vgui/ttt/icon_firegrenade.vmt")
 resource.AddFile("materials/vgui/ttt/icon_magneto_stick.vmt")
+resource.AddFile("materials/vgui/ttt/icon_smokegrenade.vmt")
 resource.AddFile("materials/vgui/ttt/icon_unarmed.vmt")
 
 -- pickup symbols